Look for plugin assets in ~/.nylas-dev when in dev mode
Nylas creates a protocol handler for nylas:// on startup. It adds a
number of paths to this, including the plugins directory. However,
when in normal mode the plugins directory is under ~/.nylas-mail, and
when in dev mode the plugins directory is under ~/.nylas-dev. The
protocol handler doesn't take this into consideration, and loads the
plugin directories under ~/.nylas-mail all the time. This means plugin
assets aren't loaded correctly when in dev mode.

This change passes the config directory to the protocol handler
creation code so it uses the correct directory.